+++ medcat/utils/ner/helpers.py
@@ -1,14 +1,15 @@
 from medcat.utils.data_utils import count_annotations
 from medcat.cdb import CDB
 
+from medcat.utils.ner.deid import deid_text as _deid_text
+from medcat.utils.decorators import deprecated
 
-def deid_text(cat, text, redact=False):
-    new_text = str(text)
-    entities = cat.get_entities(text)['entities']
-    for ent in sorted(entities.values(), key=lambda ent: ent['start'], reverse=True):
-        r = "*"*(ent['end']-ent['start']) if redact else cat.cdb.get_name(ent['cui'])
-        new_text = new_text[:ent['start']] + f'[{r}]' + new_text[ent['end']:]
-    return new_text
+
+@deprecated("API now allows creating a DeId model (medcat.utils.ner.deid.DeIdModel). "
+            "It aims to simplify the usage of DeId models. "
+            "The use of this model is encouraged over the use of this method.")
+def deid_text(*args, **kwargs) -> str:
+    return _deid_text(*args, **kwargs)
 
 
 def make_or_update_cdb(json_path, cdb=None, min_count=0):
